fixed error when reading data from file and in text tile

+++ lib/presentation/pages/dashboard/widgets/text_tile.dart
@@ -113,7 +113,7 @@ class _TextTileState extends State<TextTile> {
               return const Center(child: CircularProgressIndicator());
             }
 
-            int? percentage = result.data?["compareInNeighborhood"];
+            double? percentage = result.data?["compareInNeighborhood"];
 
             // display when all data is available
             return _getWidget(percentage: percentage?.toDouble());
